pyproject.toml
[project]
name = "market-opportunity-alert"
version = "0.1.0"
description = "your opportunity to put some money in indian stock market"
authors = [
    {name = "Vaibhav Tiwari",email = "tiwarijivaibhav@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "yfinance (>=1.1.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

